Refactor stock barcode operations into new file
