+++ test/barcode_test.dart
@@ -8,8 +8,10 @@
 import "package:flutter_test/flutter_test.dart";
 
 import "package:inventree/api.dart";
-import "package:inventree/barcode/barcode.dart";
 import "package:inventree/helpers.dart";
+
+import "package:inventree/barcode/barcode.dart";
+import "package:inventree/barcode/stock.dart";
 
 import "package:inventree/inventree/part.dart";
 import "package:inventree/inventree/stock.dart";
